// functions/src/models/person.ts
export class Person {
  id: string;
  first_name: string;
  last_name: string;
  email?: string;
  github?: string;
  bio?: string;
  google_plus?: string;
  homepage?: string;
  lanyrd?: string;
  is_speaker?: boolean;
  medium?: string;
  pictureUrl: string;
  rank: number;
  rss_feed?: string;
  short_bio?: string;
  speaker_deck?: string;
  twitter?: string;
  youtube?: string;
}